Change Trace API to only trace leaves (#3682)
Change the behavior of the Trace API to only trace leaves.  This aligns
with recent changes to CIRCT which error if annotations are applied to
aggregates which are then mandatorily split due to FIRRTL ABIs.  This
should have no practical effect on use of the Trace API only that
downstream consumers will need to work at the level of leaf types and not
on aggregates.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>